Refactor ResolvedValue into CacheableGraphqlValue
